third_party/blink/web_tests/wpt_internal/attribution-reporting/debug-key-test.sub.https.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>
const source_debug_key_promise_test = testCase =>
  attribution_reporting_promise_test(async t => {
    t.add_cleanup(() => clearCookie(testCase.cookie))

    registerAttributionSrc('Attribution-Reporting-Register-Source', {
      source_event_id :'1',
      destination: `https://{{host}}`,
      debug_key: '246',
    }, testCase.cookie);

    registerAttributionSrc('Attribution-Reporting-Register-Trigger', {
      event_trigger_data: [{trigger_data: '0'}],
    });

    const payload = await pollEventLevelReports();
    assert_equals(payload.reports.length, 1);
    const report = JSON.parse(payload.reports[0].body);

    if (testCase.keyExpected) {
      assert_equals(report.source_debug_key, '246');
    } else {
      assert_not_own_property(report, 'source_debug_key');
    }

    assert_not_own_property(report, 'trigger_debug_key');
  }, `Source: ${testCase.testName}`);

const trigger_debug_key_promise_test = testCase =>
  attribution_reporting_promise_test(async t => {
    t.add_cleanup(() => clearCookie(testCase.cookie))

    registerAttributionSrc('Attribution-Reporting-Register-Source', {
      source_event_id :'1',
      destination: `https://{{host}}`,
    });

    registerAttributionSrc('Attribution-Reporting-Register-Trigger', {
      event_trigger_data: [{trigger_data: '0'}],
      debug_key: '357',
    }, testCase.cookie);

    const payload = await pollEventLevelReports();
    assert_equals(payload.reports.length, 1);
    const report = JSON.parse(payload.reports[0].body);

    if (testCase.keyExpected) {
      assert_equals(report.trigger_debug_key, '357');
    } else {
      assert_not_own_property(report, 'trigger_debug_key');
    }

    assert_not_own_property(report, 'source_debug_key');
  }, `Trigger: ${testCase.testName}`);

const testCases = [
  {
    testName: 'Valid cookie propagates debug key.',
    cookie: 'ar_debug=1;Secure;HttpOnly;SameSite=None;Path=/',
    keyExpected: true,
  },
  {
    testName: 'Missing cookie clears debug key.',
    cookie: 'foo=1;Secure;HttpOnly;SameSite=None;Path=/',
    keyExpected: false,
  },
  {
    testName: 'Cookie without Secure clears debug key.',
    cookie: 'ar_debug=1;HttpOnly;SameSite=None;Path=/',
    keyExpected: false,
  },
  {
    testName: 'Cookie without HttpOnly clears debug key.',
    cookie: 'ar_debug=1;Secure;SameSite=None;Path=/',
    keyExpected: false,
  },
  {
    testName: 'Cookie without SameSite=None clears debug key.',
    cookie: 'ar_debug=1;Secure;HttpOnly;Path=/',
    keyExpected: false,
  },
  {
    testName: 'Cookie with narrow Path clears debug key.',
    cookie: 'ar_debug=1;Secure;HttpOnly;SameSite=None',
    keyExpected: false,
  },
];

testCases.forEach(source_debug_key_promise_test);
testCases.forEach(trigger_debug_key_promise_test);
</script>
